requirements_fedmon.txt
flwr==1.4.0
tensorflow==2.8.0
tqdm==4.64.0
prometheus-pandas==0.3.3